@@ -13,8 +13,9 @@
     "jquery": "^3.3.1",
     "moment": "^2.22.2",
     "moxios": "^0.4.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-scripts": "3.2.0",
     "react-redux": "^5.1.1",
     "react-test-renderer": "^16.6.3",
     "react_ujs": "^2.4.4",
