@@ -1,32 +1,6 @@
 import React from "react"
-import {connect} from "react-redux"
-import {fetchDate} from './actions'
 import {Container} from 'reactstrap';
 import ApplicationStatus from './components/applicationStatus'
-
-const mapStateToProps = (
-    state
-) => {
-  return {
-    ...state
-  }
-};
-
-const mapDispatchToProps = (
-    dispatch
-) => {
-  return {
-    fetchDate: () => {
-      dispatch(fetchDate())
-    }
-  }
-};
-
-const ApplicationStatusComponent = connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(ApplicationStatus);
-
 
 
 export default class LandingContainer extends React.Component {
@@ -38,7 +12,7 @@ export default class LandingContainer extends React.Component {
             <h1>Lore Character Builder</h1>
           </header>
           <section>{this.props.message}</section>
-          <ApplicationStatusComponent/>
+          <ApplicationStatus/>
         </article>
       </Container>
     )
